Домашнее задание к уроку 6 Исправление замечаний в PR

+++ src/main/java/ru/lilitweb/books/dao/BookDao.java
@@ -10,15 +10,15 @@ public interface BookDao {
     int count();
     void insert(Book book);
     void update(Book book);
-    Book getById(int id);
+    Book getById(long id);
 
     List<Book> getAll();
 
-    List<Book> getAllByGenres(int[] genres);
-    List<Book> getAllByAuthorId(int authorId);
+    List<Book> getAllByGenres(long[] genres);
+    List<Book> getAllByAuthorId(long authorId);
 
     void loadAuthors(List<Book> books, RelatedEntitiesLoader<User> usersLoader);
     void loadGenres(List<Book> books, RelatedEntitiesLoader<Genre> genresLoader);
 
-    void delete(int id);
+    void delete(long id);
 }
