+++ src/main/java/ru/lilitweb/books/dao/RelatedEntitiesLoader.java
@@ -3,5 +3,5 @@ package ru.lilitweb.books.dao;
 import java.util.List;
 
 public interface RelatedEntitiesLoader<R> {
-    List<R> getByIds(List<Integer> ids);
+    List<R> getByIds(List<Long> ids);
 }
